Cleaned up warnings
Removed unnecessary data files

+++ BH/Modules/ScreenInfo/ScreenInfo.cpp
@@ -504,7 +504,7 @@ StateCode StateCodes[] = {
 	{"FADE", 159}
 };
 
-int ExpByLevel[] = {
+long long ExpByLevel[] = {
 	0,
 	500,
 	1500,
